# YAML Parser Cache