--- a/python/src/aac/io/parser/__init__.py
+++ b/python/src/aac/io/parser/__init__.py
@@ -2,8 +2,14 @@
 
 from aac.io.parser._parse_source import parse
 from aac.io.parser._parser_error import ParserError
+from aac.io.parser._yaml import scan_yaml, parse_yaml
+from aac.io.parser._cache_manager import get_cache, reset_cache
 
 __all__ = (
     ParserError.__name__,
     parse.__name__,
+    scan_yaml.__name__,
+    parse_yaml.__name__,
+    get_cache.__name__,
+    reset_cache.__name__,
 )