--- a/python/src/aac/io/parser/_parse_source.py
+++ b/python/src/aac/io/parser/_parse_source.py
@@ -5,18 +5,20 @@ the caller with a dictionary of the content keyed by the named type.  This allow
 to find a certain type in a model by just looking for that key.
 """
 import logging
-import yaml
+from yaml import Mark, Token, StreamStartToken, StreamEndToken, DocumentStartToken
 from os import path, linesep
 from typing import Optional
-from yaml.parser import ParserError as YAMLParserError
 
 from aac.io.constants import DEFAULT_SOURCE_URI
 from aac.io.files.aac_file import AaCFile
-from aac.io.parser._parser_error import ParserError
+from aac.io.parser._cache_manager import get_cache
 from aac.io.paths import sanitize_filesystem_path
 from aac.lang.definitions.definition import Definition
 from aac.lang.definitions.lexeme import Lexeme
 from aac.lang.definitions.source_location import SourceLocation
+
+
+YAML_CACHE = get_cache()
 
 
 def parse(source: str, source_uri: Optional[str] = None) -> list[Definition]:
@@ -66,7 +68,7 @@ def _parse_str(source: str, model_content: str) -> list[Definition]:
         The AaC definitions that were built from the model contents.
     """
 
-    def mark_to_source_location(start: yaml.error.Mark, end: yaml.error.Mark) -> SourceLocation:
+    def mark_to_source_location(start: Mark, end: Mark) -> SourceLocation:
         return SourceLocation(start.line, start.column, start.index, end.column - start.column)
 
     def get_lexemes_for_definition(value_tokens, content_start, content_end) -> list[Lexeme]:
@@ -80,11 +82,11 @@ def _parse_str(source: str, model_content: str) -> list[Definition]:
     def is_token_between_locations(token, inclusive_line_start: int, inclusive_line_end: int) -> list[Lexeme]:
         return token.start_mark.line >= inclusive_line_start and token.end_mark.line <= inclusive_line_end
 
-    yaml_tokens = _scan_yaml(model_content)
-    value_tokens = [token for token in yaml_tokens if hasattr(token, "value")]
-    doc_start_token = [token for token in yaml_tokens if isinstance(token, yaml.tokens.StreamStartToken)]
-    doc_end_token = [token for token in yaml_tokens if isinstance(token, yaml.tokens.StreamEndToken)]
-    doc_segment_tokens = [token for token in yaml_tokens if isinstance(token, yaml.tokens.DocumentStartToken)]
+    yaml_tokens: list[Token] = YAML_CACHE.scan_string(model_content)
+    value_tokens: list[Token] = [token for token in yaml_tokens if hasattr(token, "value")]
+    doc_start_token: list[StreamStartToken] = [token for token in yaml_tokens if isinstance(token, StreamStartToken)]
+    doc_end_token: list[StreamEndToken] = [token for token in yaml_tokens if isinstance(token, StreamEndToken)]
+    doc_segment_tokens: list[DocumentStartToken] = [token for token in yaml_tokens if isinstance(token, DocumentStartToken)]
     doc_tokens = [*doc_start_token, *doc_segment_tokens, *doc_end_token]
 
     source_files: dict[str, AaCFile] = {}
@@ -96,14 +98,14 @@ def _parse_str(source: str, model_content: str) -> list[Definition]:
         content_start_line = start_doc_token.start_mark.line
         content_end_line = end_doc_token.end_mark.line
 
-        end_of_file_offset = 1 if isinstance(end_doc_token, yaml.tokens.StreamEndToken) else 0
+        end_of_file_offset = 1 if isinstance(end_doc_token, StreamEndToken) else 0
 
         yaml_text = linesep.join(model_content.splitlines()[content_start_line:content_end_line + end_of_file_offset])
         yaml_text += linesep
 
         if yaml_text.strip():
             definition_lexemes = get_lexemes_for_definition(value_tokens, content_start_line, content_end_line)
-            root_yaml, *_ = _parse_yaml(source, yaml_text)
+            root_yaml, *_ = YAML_CACHE.parse_string(yaml_text)
 
             if "import" in root_yaml:
                 del root_yaml["import"]
@@ -123,68 +125,6 @@ def _parse_str(source: str, model_content: str) -> list[Definition]:
             logging.debug(f"Content lines:{model_content.splitlines()}")
 
     return definitions
-
-
-def _parse_yaml(source: str, content: str) -> list[dict]:
-    """Parse content as a YAML string and return the resulting structure.
-
-    Args:
-        source (str): The source of the YAML content. Used to provide better error messages.
-        content (str): The YAML content to be parsed.
-
-    Returns:
-        The parsed YAML content.
-
-    Raises:
-        If the YAML is invalid, a ParserError is raised.
-        If the model is not a dictionary, a ParserError is raised.
-        If the model does not have (at least) a "name" field, a ParserError is raised.
-    """
-    try:
-        models = list(yaml.load_all(content, Loader=yaml.SafeLoader))
-        _error_if_not_yaml(source, content, models)
-        _error_if_not_complete(source, content, models)
-        return models
-    except YAMLParserError as error:
-        raise ParserError(source, [f"Failed to parse file, invalid YAML {error.context} {error.problem}", content])
-    except Exception as error:
-        raise ParserError(source, [f"Failed to parse file, invalid YAML {error}", content])
-
-
-def _scan_yaml(content):
-    return list(yaml.scan(content, Loader=yaml.SafeLoader))
-
-
-def _error_if_not_yaml(source, content, models):
-    """Raise a ParserError if the model is not a YAML model we can parse."""
-
-    def is_model(model):
-        """Return True if the model is further parseable."""
-        return isinstance(model, dict)
-
-    # Iterate over each model and test if it is considered a valid model.
-    if not all(map(is_model, models)):
-        raise ParserError(source, ["provided content was not YAML", content])
-
-
-def _error_if_not_complete(source, content, models):
-    """Raise a ParserError if the model is incomplete."""
-
-    def is_import(model):
-        """Return True if the model is an import declaration."""
-        type, *_ = model.keys()
-        return type == "import"
-
-    def assert_definition_has_name(model):
-        """Throws a ParserError if the definition doesn't have a name property."""
-        type, *_ = model.keys()
-        has_name = model.get(type) and model.get(type).get("name")
-        if not has_name:
-            raise ParserError(source, [f"Definition is missing field 'name': {content}"])
-
-    # Raise an error if any of the loaded YAML models are incomplete.
-    models_without_imports = list(filter(lambda m: not is_import(m), models))
-    all(map(assert_definition_has_name, models_without_imports))
 
 
 def _read_file_content(arch_file: str) -> str:
@@ -211,8 +151,7 @@ def _recurse_imports(arch_file_path: str, existing: set) -> set:
     import sequence avoiding the issue.
     """
     existing.add(arch_file_path)
-    content = _read_file_content(arch_file_path)
-    roots = _parse_yaml(arch_file_path, content)
+    roots = YAML_CACHE.parse_file(arch_file_path)
     roots_with_imports = [root for root in roots if "import" in root.keys()]
 
     for root in roots_with_imports: