--- a/python/src/aac/plugins/first_party/lsp_server/providers/semantic_tokens_provider.py
+++ b/python/src/aac/plugins/first_party/lsp_server/providers/semantic_tokens_provider.py
@@ -37,7 +37,7 @@ def _is_struct(token: Token, language_context: LanguageContext) -> bool:
 
 def _is_string(token: Token, language_context: LanguageContext) -> bool:
     """Return whether the provided token represents a string value."""
-    _ = language_context # Keeping the argument to adhere to the interface
+    _ = language_context  # Keeping the argument to adhere to the interface
     is_explicit_string = hasattr(token, "style") and token.style and token.style in "|'\""
     contains_spaces = any(map(lambda ch: ch.isspace(), token.value))
     return is_explicit_string or contains_spaces