Closes #531

* Added a caching mechanism for YAML files to reduce redundant YAML parsing

I wasn't able to catch the speed-up with profiling due to not having a particularly large modeled system (i.e. lots of yaml files), but there was some speed up in the tests. However, it looks like's there is some degree of error relying on these pipelines for repeatable timing, but there does seem to be a consistent trend in shaving ~10 seconds or so off of the test suite (recreatable with gitpod instances).

Main Branch Python Tests [https://github.com/jondavid-black/AaC/actions/runs/4474368648/jobs/7862788199](link)
3.9.0 - 112.3 secs
3.9.13 - 100.7 secs
3.10.5 - 101.4 secs

This PR's Python Tests [https://github.com/jondavid-black/AaC/actions/runs/4476603212/jobs/7867153079](link)/[https://github.com/jondavid-black/AaC/actions/runs/4476561168/jobs/7867074632](2nd link)

3.9.0 - 93.0/100.7 secs
3.9.13 - 118.7/94.5 secs
3.10.5 - 101.2/102.7 secs